test(*): extract common test setup
