+++ src/test/java/sleppynavigators/studyupbackend/common/IntegrationBaseTest.java
@@ -1,0 +1,9 @@
+package sleppynavigators.studyupbackend.common;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.test.context.ActiveProfiles;
+
+@ActiveProfiles("test")
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public class IntegrationBaseTest {
+}
