+++ src/test/java/sleppynavigators/studyupbackend/presentation/challenge/ChallengeControllerTest.java
@@ -15,14 +15,11 @@ import java.time.LocalDateTime;
 import java.util.List;
 import java.util.Objects;
 import org.apache.http.HttpStatus;
-import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.test.web.server.LocalServerPort;
-import org.springframework.test.context.ActiveProfiles;
+import sleppynavigators.studyupbackend.common.RestAssuredBaseTest;
 import sleppynavigators.studyupbackend.domain.authentication.token.AccessToken;
 import sleppynavigators.studyupbackend.domain.authentication.token.AccessTokenProperties;
 import sleppynavigators.studyupbackend.domain.challenge.Challenge;
@@ -36,12 +33,9 @@ import sleppynavigators.studyupbackend.presentation.challenge.dto.request.Challe
 import sleppynavigators.studyupbackend.presentation.challenge.dto.request.ChallengeCreationRequest.TaskRequest;
 import sleppynavigators.studyupbackend.presentation.challenge.dto.request.TaskCertificationRequest;
 import sleppynavigators.studyupbackend.presentation.challenge.dto.response.TaskCertificationDTO;
-import sleppynavigators.studyupbackend.presentation.common.DatabaseCleaner;
 
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-@ActiveProfiles("test")
 @DisplayName("ChallengeController API 테스트")
-public class ChallengeControllerTest {
+public class ChallengeControllerTest extends RestAssuredBaseTest {
 
     @Autowired
     private GroupRepository groupRepository;
@@ -55,12 +49,6 @@ public class ChallengeControllerTest {
     @Autowired
     private AccessTokenProperties accessTokenProperties;
 
-    @Autowired
-    private DatabaseCleaner databaseCleaner;
-
-    @LocalServerPort
-    private int port;
-
     private User currentUser;
 
     @BeforeEach
@@ -72,17 +60,10 @@ public class ChallengeControllerTest {
                 new AccessToken(currentUser.getId(), userProfile, List.of("profile"), accessTokenProperties);
         String bearerToken = "Bearer " + accessToken.serialize(accessTokenProperties);
 
-        RestAssured.port = port;
         RestAssured.requestSpecification = new RequestSpecBuilder()
                 .setContentType(ContentType.JSON)
                 .addHeader("Authorization", bearerToken)
                 .build();
-    }
-
-    @AfterEach
-    void tearDown() {
-        databaseCleaner.execute();
-        RestAssured.reset();
     }
 
     @Test
